Step 3: Add a bespoke DittoProcessorAttribute
This step adds in a bespoke DittoProcessorAttribute, while it's unlikely you would want to return the length of the string for your title property this is an example of how a Processor can modify the type.

Ditto's engine handles turning an int to string after it leaves the StringLengthProcessor.

Added in new bespoke Processor, StringLengthProcessor, to HomePageViewModel.