Step 4: Add in a bespoke with a custom context
Finally we're wiring in a bespoke context which allows a number to be passed to a processor.

The number comes from the controller but it could come from an appsetting, database or anywhere. The idea is to allow the processor to not care where the data comes from but it knows it may have some data to use.

This step adds in the new bespoke processor which accesses the context through the  DittoProcessorMetaDataAttribute.

Added context to controller & decorated the view model with the processor attribute.